requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
PyYAML==6.0.1

# Database
sqlite3
sqlalchemy==2.0.23
alembic==1.13.1

# Slack integration
slack-sdk==3.26.1
slack-bolt==1.18.1

# GitHub integration
PyGithub==1.59.1
requests==2.31.0

# AI/LLM integration
openai==1.3.7
anthropic==0.7.8
langchain==0.0.350
langchain-openai==0.0.2

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
schedule==1.2.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1 